feat: add rest client to encapsulate http
