+++ Client/InfluxDBClient.cs
@@ -1,12 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
-using System.IO;
-using System.Linq;
 using System.Net.Http;
-using System.Net.Http.Headers;
-using System.Runtime.Serialization;
-using System.Runtime.Serialization.Json;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -58,15 +53,10 @@ namespace InfluxDB3.Client
 
     public class InfluxDBClient : IInfluxDBClient
     {
-        private static readonly string[] ErrorHeaders =
-            { "X-Platform-Error-Code", "X-Influx-Error", "X-InfluxDb-Error" };
-
-        private readonly InfluxDBClientConfigs _configs;
-
         private bool _disposed;
 
         private readonly FlightSqlClient _flightSqlClient;
-        private readonly HttpClient _httpClient;
+        private readonly RestClient _restClient;
 
         /// <summary>
         /// This class provides an interface for interacting with an InfluxDB server,
@@ -91,21 +81,13 @@ namespace InfluxDB3.Client
         /// <param name="configs">The configuration of the client.</param>
         public InfluxDBClient(InfluxDBClientConfigs configs)
         {
-            _configs = configs ??
-                       throw new ArgumentException("The configuration of the client has to be defined.");
-
-            _flightSqlClient = new FlightSqlClient(host: _configs.Host);
-            _httpClient = new HttpClient(new HttpClientHandler
+            if (configs is null)
             {
-                AllowAutoRedirect = _configs.AllowHttpRedirects
-            });
-
-            _httpClient.Timeout = _configs.Timeout;
-            _httpClient.DefaultRequestHeaders.UserAgent.ParseAdd($"influxdb3-csharp/{AssemblyHelper.GetVersion()}");
-            if (!string.IsNullOrEmpty(configs.Token))
-            {
-                _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Token", configs.Token);
+                throw new ArgumentException("The configuration of the client has to be defined.");
             }
+
+            _flightSqlClient = new FlightSqlClient(host: configs.Host);
+            _restClient = new RestClient(configs: configs);
         }
 
         /// <summary>
@@ -178,66 +160,15 @@ namespace InfluxDB3.Client
                 return;
             }
 
-            var request = new HttpRequestMessage
-            {
-                Method = HttpMethod.Post,
-                RequestUri = new Uri($"{_configs.Host}api/v2/write"),
-                Content = new StringContent(sb.ToString(), Encoding.UTF8, "text/plain")
-            };
+            var content = new StringContent(sb.ToString(), Encoding.UTF8, "text/plain");
 
-            request.Headers.Add("Accept", "application/json");
-
-            string? message = null;
-            var result = await _httpClient.SendAsync(request, cancellationToken).ConfigureAwait(false);
-            if (!result.IsSuccessStatusCode)
-            {
-                var content = await result.Content.ReadAsStringAsync().ConfigureAwait(true);
-                // error message in body
-                if (!string.IsNullOrEmpty(content))
-                {
-                    using var memoryStream = new MemoryStream(Encoding.UTF8.GetBytes(content));
-                    try
-                    {
-                        if (new DataContractJsonSerializer(typeof(ErrorBody)).ReadObject(memoryStream) is ErrorBody errorBody)
-                        {
-                            message = errorBody.Message;
-                        }
-                    }
-                    catch (SerializationException se)
-                    {
-                        Debug.WriteLine($"Cannot parse error response as JSON: {content}. {se}");
-                    }
-                }
-
-                // from header
-                if (string.IsNullOrEmpty(message))
-                {
-                    message = result.Headers?
-                        .Where(header => ErrorHeaders.Contains(header.Key, StringComparer.OrdinalIgnoreCase))
-                        .Select(header => header.Value.FirstOrDefault()?.ToString())
-                        .FirstOrDefault();
-                }
-
-                // whole body
-                if (string.IsNullOrEmpty(message))
-                {
-                    message = content;
-                }
-
-                // reason
-                if (string.IsNullOrEmpty(message))
-                {
-                    message = result.ReasonPhrase;
-                }
-
-                throw new InfluxDBApiException($"Cannot write data to InfluxDB due: {message}", result);
-            }
+            await _restClient.Request("api/v2/write", HttpMethod.Post, content, cancellationToken);
         }
 
         public void Dispose()
         {
             _flightSqlClient.Dispose();
-            _httpClient.Dispose();
+            _restClient.Dispose();
             _disposed = true;
         }
 
@@ -271,11 +202,5 @@ namespace InfluxDB3.Client
 
             return sb;
         }
-    }
-
-    [DataContract]
-    internal class ErrorBody
-    {
-        [DataMember(Name = "error")] public string? Message { get; set; }
     }
 }
